Refactored QuantME transformation
