+++ components/bpmn-q/.eslintrc.js
@@ -22,5 +22,6 @@ module.exports = {
     "rules": {
         "semi": [2, "always"],
         "react/prop-types": "off",
+        "no-prototype-builtins": "off",
     }
 };
