+++ components/bpmn-q/modeler-component/extensions/quantme/replacement/InputOutputHandler.js
@@ -9,8 +9,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { getPropertiesToCopy } from '../utilities/Utilities';
 import { QUANTME_ATTRIBUTES } from '../Constants';
+import {getPropertiesToCopy} from '../../../common/util/TransformationUtilities';
 
 /**
  * Add attributes of a replaced QuantME task to the replacing workflow fragment to enable its configuration based of the attribute values.
